requirements.txt
# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7

# Data Processing and Analysis
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# API Documentation
Flask-Swagger-UI==4.11.1

# Environment and Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.4.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Development Tools
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.25.2
